RaspberryPI/requirements.txt
python-rtmidi
board
adafruit-circuitpython-neopixel
mido
flask
